Upgrade for Terraform v0.13
Why these changes are being introduced:
In order to upgrade the mitlib-terraform to work with Terraform
v0.13, we need to upgrade all of the modules to support v0.13.

How this addresses that need:
* The only 0.13-related change to this code is updating the
required_version constraint in the `terraform { }` block.

There are only changes being introduced as part of this upgrade:
* .travis.yml is removed to disconnect this repo from TravisCI
* a GitHub Action is introduced to replace Travis (for checking
Terraform syntax and validating the code)
* a modified version of our standard Terraform PR template is now
included in this repository.

Relevant Jira pages:
* There is no Jira ticket for this work, but there is a InfraEng
private KB article outlining our Terraform v0.13 upgrade process:
https://mitlibraries.atlassian.net/l/c/fYft9yi1